Fisher exact test alternative="greater"
This is a major bugfix. In some cases, even genes with num_goterms_product_process = 0 were labelled as statistically significant for the regulation of a given process. The problem was that the fisher exact test uses a two-sided approach by default, thus it also labels underrepresented samples as statistically significant, which was wrong. The bugfix was achieved by the parameter alternative="greater" when calling scipy.stats.fisher_exact.