GitHub CI and CD workflows reverted to Python versions 3.10 and 3.11
This is a follow-up of the pyzmq error, which prevented GOReverseLookup from being pushed to pypi. The error happened when TestPypi and Pypi began using Python 3.12 for test installs of the GOReverseLookup package. The last successful Pypi push was made with GOReverseLookup 3.0.10, which was ran using Python 3.11.x (see Github actions history and search for keyword "python").