Indirect annotations parent/child modularity and depth
Choosing whether to take GO term parents or children of directly annotated GO terms (to a gene) is now supported. This can be set as the second parameter of the 'include_indirect_annotations' setting, by setting either a "p" (parent) or "c" (child). Example: 'include_indirect_annotations True p'.

For indirect annotations, a max depth setting was implemented, which takes only 'max_depth' indirect annotations, discarding the rest. This is to eliminate GO terms which are high in the Gene Ontology hierarchy (such as "regulation of a biological process") from being included in the gene scoring process. This can be set via the 'indirect_annotations_max_depth' setting. Setting it to -1 means to take the full depth of the hierarchy.